Address Sphinx 2 compatibility issues
The main change here is that we load information into the template so
that CSS can make decisions about what element hierarchy it can expect,
instead of covering both html4 and html5 writer with a single set of
rules. I also fixed the tox config so that it generates docs for
testing.

This is also in part going to revert the changes in #785. Instead of
maintaining a list of `dl` classes to apply the domain styles to, this
PR instead maintains a list of definition list classes to *not* apply
the domain definition list styling too. There are too many possibilities
to cover with an include list.